refactor: extract LocalDataDeletionTypes.swift from LocalDataDeletionController.swift
Splits the `LocalDataDeletionOutcome` struct + `LocalDataDeletionResult`
enum out of LocalDataDeletionController.swift into a dedicated Types
file. That leaves LocalDataDeletionController.swift holding only the
controller class (matching its filename).

Byte-identical move of lines 3-23. Both files import Foundation.

LocalDataDeletionController.swift: 70 → 49 lines. Tests: 667.
